backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
SQLAlchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
pydantic>=1.8.0
pyvisa>=1.11.0
requests>=2.26.0
bcrypt>=3.2.0
python-jose[cryptography]>=3.3.0
celery[redis]>=5.1.0
websockets>=10.0
pytest>=6.2.0
pytest-asyncio>=0.15.0
httpx>=0.18.0
pytest-cov>=2.12.0
flake8>=3.9.0
black>=21.7b0
